Correct phone validation
The phone input is not required, as such it passed the validity test even when it was empty. We want to keep that but not display correct or incorrect next to it since it's confusing.
Made a different event listener for the phone, checked if the input has value and only display if it does.